HIVE-2848 [jira] Add validation to HiveConf ConfVars
Summary:
DPAL-970 Add validation to HiveConf ConfVars

It would be good if we could add some validation to ConfVars in the HiveConf
when they are set.

At least if a variable is supposed to be boolean, numeric, or strict/nonstrict,
it shouldn't be too difficult to check for these.

Test Plan: EMPTY

Reviewers: JIRA

Differential Revision: https://reviews.facebook.net/D2313