HIVE-4152 [jira] Lock owner should be allowed to acquire lock again on EmbeddedLockManager
Summary:
HIVE-4152 Lock owner should be allowed to acquire lock again on EmbeddedLockManager

If a query owns lock on a path, it should be allowed to acquire lock on it again with success. Current EmbeddedLockManager returns null in this situation and makes some query to wait infinitely.

Test Plan: EMPTY

Reviewers: JIRA

Differential Revision: https://reviews.facebook.net/D9303